% Generated by roxygen2 (4.0.2): do not edit by hand
\name{acc}
\alias{acc}
\title{Absolute Average Accuracy}
\usage{
acc(y, x)
}
\arguments{
\item{y}{dataset 1}

\item{x}{dataset 2}
}
\value{
numeric
}
\description{
This function will calculate the absolute average accuracy of compartive datasets
}
\examples{
\dontrun{
acc(y,x)
}
}

