% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{age.years}
\alias{age.years}
\title{Age Years}
\usage{
age.years(earlier, later)
}
\arguments{
\item{earlier}{first date}

\item{later}{second date}
}
\value{
integer
}
\description{
This function for calculating age with two dates
}
\examples{
\dontrun{
x <- as.Date("2000-02-29")
y <- as.Date("2004-02-28")
age.years(x, y)
}
}
