% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ATR210}
\alias{ATR210}
\title{ATR Bands}
\usage{
ATR210(prices, innervalue = 1.5, outervalue = 2, length = 10)
}
\arguments{
\item{innervalue}{numeric factor for inner bands}

\item{outervalue}{numeric factor for outer bands}

\item{length}{numeric value to measure ATR}
}
\description{
ATR Bands
}
