% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLB.R
\name{bbref_leaderboard}
\alias{bbref_leaderboard}
\title{Scrape Pitching or Hitting data from Baseball Reference}
\usage{
bbref_leaderboard(bat_pitch, t1 = NULL, t2 = NULL, n = 30)
}
\arguments{
\item{bat_pitch}{either 'bat' or 'pit'}

\item{t1}{First date data should be scraped from. Should take the form "YEAR-DAY-MONTH"}

\item{t2}{Second date data should be scraped from. Should take the form "YEAR-DAY-MONTH"}

\item{n}{use number of days prior from today}
}
\description{
Original code from Bill Petti and modified as needed
}
\examples{
bb_p<-bbref_leaderboard('pit',"2015-05-10", "2015-06-20")
bb_h<-bbref_leaderboard('bat',n=30) # current year; last 30 days from today
}
