% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{btween}
\alias{btween}
\title{Between}
\usage{
btween(x, low, high, ineq = F)
}
\arguments{
\item{x}{numeric}

\item{low}{numeric}

\item{high}{numeric}

\item{ineq}{boolean}
}
\value{
logical
}
\description{
This function mimics the SQL between clause and returns a logical indx
}
\examples{
\dontrun{
btween( 1:10, 5.5, 6.5 )
}
}
