% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{buy.sell}
\alias{buy.sell}
\title{Buy/Sell Indicator}
\usage{
buy.sell(x)
}
\arguments{
\item{x}{numeric}
}
\value{
numeric
}
\description{
This function will provide one buy and one sell indicator for the given
series
}
\examples{
\dontrun{
buy.sell(x)
}
}

