% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cnt}
\alias{cnt}
\title{Count number of non NA elements}
\usage{
cnt(x, side = 2)
}
\arguments{
\item{x}{data matrix}

\item{side}{margin along which to count}
}
\value{
counts
}
\description{
This function will count number of non NA elements in the given matrix
}
\examples{
\dontrun{
cnt(matrix(c(1,NA,2,3),2,2))
}
}

