% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLB.R
\name{ESPN_eligibility}
\alias{ESPN_eligibility}
\title{Returns a full list of MLB players eligibility 
 for hitters or pitchers for ESPN Standard leagues'
 Retreives primary position only}
\usage{
ESPN_eligibility(bat_pitch)
}
\arguments{
\item{bat_pitch}{either 'bat' or 'pit'}
}
\value{
data frame
}
\description{
Returns a full list of MLB players eligibility 
 for hitters or pitchers for ESPN Standard leagues'
 Retreives primary position only
}

