% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLB.R
\name{ESPN_proj}
\alias{ESPN_proj}
\title{Returns projections for hitters or pitchers for a standard 5x5 ROTO leagues}
\usage{
ESPN_proj(bat_pitch, leagueID = 0)
}
\arguments{
\item{bat_pitch}{either 'bat' or 'pit'}

\item{leagueID}{ESPN league ID}
}
\description{
Returns projections for hitters or pitchers for a standard 5x5 ROTO leagues
}
\examples{
ESPN_hitters<-ESPN_proj('bat')

}

