\name{extendForecast}
\alias{extendForecast}
\title{Adding future dates to a forecast}
\usage{
  extendForecast(dates, forecast)
}
\arguments{
  \item{dates}{the dates of the historical values}

  \item{forecast}{a vector of forecasted values}
}
\description{
  Adding future dates to a forecast
}
\examples{
##NULL
}
\keyword{dates}
\keyword{forecast}

