% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLB.R
\name{fangraphs_leaderboard}
\alias{fangraphs_leaderboard}
\title{Scrape Batter Leaderboards from FanGraphs.com}
\usage{
fangraphs_leaderboard(bat_pitch, yearfrom = NULL, yearto = NULL,
  qual = NULL, split = NULL)
}
\arguments{
\item{bat_pitch}{either 'bat' or 'pit'}

\item{yearfrom}{First season for which you want data.}

\item{yearto}{Last season for which you want data. If multiple years selected, data returned will be aggregate data for the date range. If yearto = yearfrom, function will return single-season data.}

\item{qual}{Whether you want only batters that qualified in a given season, or the minimum number of plate appearances for inclusion. If you only want qualified hitters, use qual. If a minimumm number of plate appearaces, use the number desired.}

\item{split}{'0' - full season '1' - 7 days '2' - 15 days 3' - 30 days '30' - 1st half'31' - 2nd half '15' - Home '16' - Away '13' - vs LHP or LHB '14' - vs RHP or RHB}
}
\value{
data frame
}
\description{
This function allows you to scrape all leaderboard statistics from FanGraphs.com.
}
\examples{
fgh<-fangraphs_leaderboard('bat',2013,2015, 200, 0)

}

