\name{getQuote.google}
\alias{getQuote.google}
\title{Get the latest prices from the Google finance:}
\usage{
  getQuote.google(tickers)
}
\arguments{
  \item{tickers}{symbols}
}
\description{
  Get the latest prices from the Google finance:
}
\examples{
\dontrun{
getQuote.google(spl('MSFT,AAPL,IBM'))
}
}

