\name{getSymbols.PI}
\alias{getSymbols.PI}
\title{Download historical prices from Pi Trading - Free Market Data}
\usage{
  getSymbols.PI(Symbols, env = .GlobalEnv,
    auto.assign = TRUE, download = TRUE)
}
\arguments{
  \item{Symbols}{tickers}

  \item{env}{envelope}

  \item{auto.assign}{assignment}

  \item{download}{downloaded}
}
\description{
  Download historical prices from Pi Trading - Free Market
  Data
}
\examples{
\dontrun{
getSymbols.PI()
}
}

