% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLB.R
\name{Heatmaps_leaderboard}
\alias{Heatmaps_leaderboard}
\title{Scrape Baseball Heatmaps leadersboards}
\usage{
Heatmaps_leaderboard(bat_pitch = "bat")
}
\arguments{
\item{bat_pitch}{either 'bat' or 'pit'}
}
\description{
Scrape Baseball Heatmaps leadersboards
}
\details{
Limitations: Cannot change year. Always current
}
\examples{
heatmaps<-Heatmaps_leaderboard('bat')
}
