% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{indx.xts}
\alias{indx.xts}
\title{Fast alternative to indx() function for \code{\link{xts}} object}
\usage{
indx.xts(x)
}
\arguments{
\item{x}{\code{\link{xts}} object}
}
\value{
dates
}
\description{
NOTE indx.xts is the same name as the indx function in the XTS package
}
\examples{
\dontrun{ 
indx.xts(make.xts(1:101,seq(Sys.Date()-100, Sys.Date(), 1)))
}
}

