% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.lines}
\alias{plot.lines}
\title{plot.lines}
\usage{
plot.lines(y, type = "l", col = par("col"), ...)
}
\arguments{
\item{y}{xts object to plot}

\item{type}{plot type}

\item{col}{color}

\item{...}{other parameters to lines}
}
\description{
plot.lines
}
