\name{prettyR}
\alias{prettyR}
\title{Formatting R code snippets}
\usage{
  prettyR(file)
}
\arguments{
  \item{file}{filepath of R script}
}
\description{
  Using code borrowed from Distantobserver who wrote a
  function that replaces the copy-and-paste part of the
  PrettyR webtool
}
\examples{
## NULL
}
\keyword{Formatting}
\keyword{PrettyR}

