% Generated by roxygen2 (4.0.2): do not edit by hand
\name{remove.outliers}
\alias{remove.outliers}
\title{Remove Outliers from a dataset}
\usage{
remove.outliers(x, na.rm = TRUE, ...)
}
\arguments{
\item{x}{object}

\item{na.rm}{boolean}
}
\value{
object
}
\description{
This function removes outliers from a dataset
credit for this source code goes to aL3xa on StackOverflow
}
\examples{
\dontrun{
set.seed(1)
x <- rnorm(100)
x <- c(-10, x, 10)
y <- remove_outliers(x)
par(mfrow = c(1, 2))
boxplot(x)
boxplot(y)
}
}

