% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{repCol}
\alias{repCol}
\title{Repeat Columns}
\usage{
repCol(m, nc)
}
\arguments{
\item{m}{vector (column)}

\item{nc}{number of copies along columns}
}
\value{
new matrix
}
\description{
Repeat Columns
}
\examples{
\dontrun{
matrix(1:5, nr=5, nc=3, byrow=F)
repCol(1:5, 3)
}
}
