% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{SCTR}
\alias{SCTR}
\title{THIS STUDY IS A REPLICATION OF STOCKCHARTS TECHNICAL RANKING
For more info:
http://stockcharts.com/school/doku.php?id=chart_school:technical_indicators:sctr}
\usage{
SCTR(x)
}
\arguments{
\item{x}{vector or xts}
}
\description{
THIS STUDY IS A REPLICATION OF STOCKCHARTS TECHNICAL RANKING
For more info:
http://stockcharts.com/school/doku.php?id=chart_school:technical_indicators:sctr
}

