% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sector.spdr.components}
\alias{sector.spdr.components}
\title{Get Sector SPDR Components}
\usage{
sector.spdr.components(sector.etf = "XLE")
}
\arguments{
\item{sector.etf}{SPDR symbol}
}
\value{
tickers
}
\description{
Get Sector SPDR Components
}
\examples{
\dontrun{
sector.spdr.components()
}
}

