% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLB.R
\name{statcast_leaderboard}
\alias{statcast_leaderboard}
\title{Scrape Batted Ball Distance/Velocity Data from Baseball Savant / MLB Statcast}
\usage{
statcast_leaderboard(bat_pitch = "bat", year = NULL, from.date = NULL,
  qual = 20)
}
\arguments{
\item{bat_pitch}{either 'bat' or 'pit'}

\item{year}{YYYY number format}

\item{from.date}{YYYY-MM-DD format}

\item{qual}{Number of ABs that meets the qualification}
}
\value{
data frame
}
\description{
This function allows you to scrape all leaderboard statistics from the Baseball Savant batted ball data leaderboard
}
\examples{
statcast<-statcast_leaderboard('bat', 2016, "2016-04-10", 20)
}

