% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str_detach}
\alias{str_detach}
\title{Detach prefixes or suffixes}
\usage{
str_detach(x, sep = ",", side = "left", greedy = TRUE)
}
\arguments{
\item{...}{arguments to \code{str_filter}}
}
\description{
Detach prefixes or suffixes
}
\examples{
ex = c("a", "a, b", "a, b, c", "a, b, c, d")
str_detach(ex, side = "left"  , greedy = TRUE)
str_detach(ex, side = "right" , greedy = TRUE)
str_detach(ex, side = "left"  , greedy = FALSE)
str_detach(ex, side = "right" , greedy = FALSE)
}
\references{
http://www.r-bloggers.com/string-manipulations-on-full-names/
}
