% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str_filter}
\alias{str_filter}
\title{Remove prefixes or suffixes}
\usage{
str_filter(x, sep = ",", side = "left", greedy = TRUE)
}
\arguments{
\item{sep}{character separator, defaults to ","}

\item{side}{where to look, "left" (default) or "right"}

\item{greedy}{defaults to TRUE}
}
\description{
Remove prefixes or suffixes
}
\examples{
ex = c("a", "a, b", "a, b, c", "a, b, c, d")
str_filter(ex, side = "left"  , greedy = TRUE)
str_filter(ex, side = "right" , greedy = TRUE)
str_filter(ex, side = "left"  , greedy = FALSE)
str_filter(ex, side = "right" , greedy = FALSE)
}
\references{
#' @references http://www.r-bloggers.com/string-manipulations-on-full-names/
}

