\name{unadj.rsquared}
\alias{unadj.rsquared}
\title{Unadjusted R Squared}
\usage{
  unadj.rsquared(fit)
}
\arguments{
  \item{fit}{lm}
}
\value{
  numeric
}
\description{
  This function will calculate unadjusted R squared value
  which explains the model more accurately, because the
  intercept is factored back into the model
}
\examples{
\dontrun{
unadj.rsquared(fit)
}
}

