% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLB.R
\name{WarGames_ESPN_proj}
\alias{WarGames_ESPN_proj}
\title{Returns projections for hitters pr pitchers for WAR Games ESPN FB league (ONLY)
because column headings will be different for other leagues!!}
\usage{
WarGames_ESPN_proj(bat_pitch, leagueID = 86607)
}
\arguments{
\item{bat_pitch}{either 'bat' or 'pit'}

\item{leagueID}{ESPN league ID}
}
\description{
Returns projections for hitters pr pitchers for WAR Games ESPN FB league (ONLY)
because column headings will be different for other leagues!!
}
\examples{
WARGames.Hitters<-WarGames_ESPN_proj('bat',86607)
}

