% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{date.dayofweek}
\alias{date.dayofweek}
\alias{date.day}
\alias{date.week}
\alias{date.month}
\alias{date.year}
\title{Dates Functions}
\usage{
date.dayofweek(dates)

date.day(dates)

date.week(dates)

date.month(dates)

date.year(dates)
}
\arguments{
\item{dates}{collection of dates}
}
\value{
transformed dates
}
\description{
Dates Functions
}
\examples{
\dontrun{
date.dayofweek(Sys.Date())
}
}
