% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{repRow}
\alias{repRow}
\title{Repeat Rows}
\usage{
repRow(n, nr)
}
\arguments{
\item{nr}{number of copies along rows}

\item{m}{vector (row)}
}
\value{
new matrix
}
\description{
Repeat Rows
}
\examples{
\dontrun{
matrix(1:3, nr=5, nc=3, byrow=T)
repRow(1:3, 5)
}
}

