% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLB.R
\name{savant_leaderboard}
\alias{savant_leaderboard}
\title{Stcrape Batted Ball Distance/Velocity Data from Baseball Savan}
\usage{
savant_leaderboard(bat_pitch, qual = 20)
}
\arguments{
\item{bat_pitch}{either 'bat' or 'pit'}

\item{qual}{Number of ABs that meets the qualification}
}
\value{
data frame
}
\description{
This function allows you to scrape all leaderboard statistics from the Baseball Savant batted ball data leaderboard
}
\examples{
savant<-savant_leaderboard(20)
}

