% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{SCTR}
\alias{SCTR}
\title{StockCharts Technical Rank}
\usage{
SCTR(x)
}
\arguments{
\item{x}{vector or xts}
}
\description{
StockCharts Technical Rank
}
\references{
\url{http://stockcharts.com/school/doku.php?id=chart_school:technical_indicators:sctr}
}

